se suben ultimos cambios

+++ components/tempHumidity/include/tempHumidity.h
@@ -8,8 +8,8 @@ typedef struct
     float humidity1;
     float temperature2;
     float humidity2;
-    float humGroud1;
-    float humGroud2;
+    int humGroud1;
+    int humGroud2;
 } tempHumidity_t;
 
 char* convert_to_json_string(tempHumidity_t* data);
